Added antialiasing.  Changed Scene.cpp to use antialiasing as described on page 311 of tiger.  Also, commented out instances where we used new and instead just used stack memory to make the objects.  This removed memory leaks.

+++ rayTracer/Sample.h
@@ -3,8 +3,8 @@
 class Sample
 {
 public:
-	int x, y;
+	float x, y;
 	Sample(void);
-	Sample(int posx, int posy);
+	Sample(float posx, float posy);
 	~Sample(void);
 };
