+++ rayTracer/Sampler.cpp
@@ -9,6 +9,7 @@ Sampler::Sampler(int sizex, int sizey)
 	tenPercent = totalPixels / 10;
 	level = 1;
 	toMilestone = 0;
+	
 }
 
 Sampler::~Sampler(void)
@@ -17,8 +18,9 @@ Sampler::~Sampler(void)
 
 bool Sampler::getSample(Sample *s) {
 	if(curr < totalPixels) {
-		s->x = curr % mySizex;
-		s->y = curr / mySizex;
+		s->x = (float) (curr % mySizex);
+		s->y = (float) (curr / mySizex);
+		
 		curr ++;
 		printProgress();
 		return true;
